Implement a few partial use cases

+++ se.chalmers.cse.mdsd1617.group13/src/se/chalmers/cse/mdsd1617/group13/hotelsystem/Bill.java
@@ -57,12 +57,12 @@ public interface Bill extends EObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Price</em>' attribute.
-	 * @see #setPrice(int)
+	 * @see #setPrice(double)
 	 * @see se.chalmers.cse.mdsd1617.group13.hotelsystem.HotelsystemPackage#getBill_Price()
 	 * @model required="true" ordered="false"
 	 * @generated
 	 */
-	int getPrice();
+	double getPrice();
 
 	/**
 	 * Sets the value of the '{@link se.chalmers.cse.mdsd1617.group13.hotelsystem.Bill#getPrice <em>Price</em>}' attribute.
@@ -72,6 +72,6 @@ public interface Bill extends EObject {
 	 * @see #getPrice()
 	 * @generated
 	 */
-	void setPrice(int value);
+	void setPrice(double value);
 
 } // Bill
